updated footer and designed error page
