+++ Frontend/src/components/pages/DeletePost.jsx
