+++ Frontend/src/components/pages/Home.jsx
@@ -1,0 +1,15 @@
+import React from 'react'
+import '../../css/Home.css'
+import Header from '../Header'
+import Footer from '../Footer'
+
+const Home = () => {
+  return (
+    <div>
+      <Header/>
+      <Footer/>
+    </div>
+  )
+}
+
+export default Home
