+++ Frontend/src/components/pages/Logout.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+const Logout = () => {
+  return (
+    <div>
+      
+      <div>
+      </div>
+      
+    </div>
+  );
+};
+
+export default Logout;
