+++ Frontend/src/components/pages/Register.jsx
