--- Frontend/src/pages/ErrorPage.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const ErrorPage = () => {
-  return (
-    <div>
-      
-    </div>
-  )
-}
-
-export default ErrorPage
