--- Frontend/src/pages/Home.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const Home = () => {
-  return (
-    <div>
-      
-    </div>
-  )
-}
-
-export default Home
