--- Frontend/src/pages/Logout.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const Logout = () => {
-  return (
-    <div>
-      
-    </div>
-  )
-}
-
-export default Logout
